requirements.txt
discord.py>=2.3.2
deep-translator>=1.9.1
python-dotenv
sqlite3
langdetect